feat: sync events by address and fetch all logs
Generally when filtering by address we want to reduce the amount of events we process/store in the database. However most of the time we are not only interested in the events only emitted by a particular contract, but also in the other events in the transaction. For example, if we want to fetch all the events for particular NFT I would argue that a sale on opensea is also event related to the NFT, but since the orderFullfilled event from seaport is not emitted by the contract address we filter on, we end up ommitting the events.
I’ve added a fetchAllTxLogs param to the syncDetail when filtering by address. This will basically perform additional eth_getTransactionReceipt calls for every transactionHash. In some situatioons those calls might be redundant, but most of the time if we are dealing with sales on marketplaces we want the additional events.