feat: realtime sync only known collections
Currently when syncing realtime data we basically store everything that resembles an NFT or comes from a known marketplace.
In our case we are only interested in keeping up with info on just a few NFT collections and we are not interested in everything that happens on chain.

This commit introduces an opt-in options for syncing only knownCollections. If the user sets CATCHUP_FOR_ONLY_KNOWN_COLLECTIONS to true, then we’ll get the contract addresses of the collections that have been already synced (through backfill for example) and we’ll look for logs only for those collections.